requirements.txt
Pillow==5.0.0
wheel==0.24.0
